Improve parallelization [NSETM-1435]
Execute strategies sequentially, but parallelize tasks inside the strategy.

Previously, s2f-recipe was generating parallel tasks for all the
strategies, then executing them in parallel.
However, in this way it was difficult to add tasks that should be processed
in parallel in a different way (for example in estimate_bouton_reduction
the tasks should be split by gid, instead of pathway, and the results
should be combined differently).

List of strategies:
- estimate_bouton_reduction: single task executing subtasks in subprocesses
- estimate_individual_bouton_reduction: parallel tasks
- estimate_syns_con: parallel tasks
- existing_recipe: single task
- experimental_syns_con: single task
- generalized_cv: single task
- override_mtype: single task